+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- [#127](https://github.com/os2display/display-client/pull/127)
+  - Migrate from CRA to Vite
+  - Upgrade node from 14 -> 20
+  - Winston logging to Pino logging
+  - Set infrastructure node version to 20.
+  - Fixed base path to /client.
+
 ## [2.0.4] - 2024-08-14
 
 - [#128](https://github.com/os2display/display-client/pull/128)